#466 refactor create project form into container component
