+++ src/components/projects/create-project-form/create-project-form.test.tsx
@@ -4,54 +4,19 @@
  */
 
 import { render, screen } from "../../../test-support/test-utils";
-import CreateProjectForm from "./create-project-form";
+import CreateProjectFormView from "./create-project-form/create-project-form";
 
 /**
  * Sets up the component under test with the desired values and renders it.
  */
 const renderComponent = () => {
-  return render(<CreateProjectForm />);
+  return render(<CreateProjectFormView />);
 };
 
 describe('create project form test suite', () => {
-  it('renders title', () => {
+  it('render view component', () => {
     renderComponent();
 
-    expect(screen.queryByText('New Project')).toBeInTheDocument();
-  });
-
-  it('renders project name form input', () => {
-    renderComponent();
-
-    expect(screen.getByLabelText('Project Name')).toBeInTheDocument();
-    expect(screen.getByPlaceholderText('Enter project name...')).toBeInTheDocument();
-  });
-
-  it('renders car number form input', () => {
-    renderComponent();
-
-    expect(screen.getByLabelText('Car Number')).toBeInTheDocument();
-    expect(screen.getByPlaceholderText('Enter car number...')).toBeInTheDocument();
-  });
-
-  it('renders change request id form input', () => {
-    renderComponent();
-
-    expect(screen.getByLabelText('Change Request ID')).toBeInTheDocument();
-    expect(screen.getByPlaceholderText('Enter change request ID...')).toBeInTheDocument();
-  });
-
-  it('renders project summary form input', () => {
-    renderComponent();
-
-    expect(screen.getByLabelText('Project Summary')).toBeInTheDocument();
-    expect(screen.getByPlaceholderText('Enter summary...')).toBeInTheDocument();
-  });
-
-  it('renders buttons', () => {
-    renderComponent();
-
-    expect(screen.getByText('Create')).toBeInTheDocument();
-    expect(screen.getByText('Cancel')).toBeInTheDocument();
-  });
+    expect(screen.getByText('New Project')).toBeInTheDocument();
+  })
 });
