+++ src/components/projects/create-project-form/create-project-form/create-project-form.module.css
@@ -2,5 +2,5 @@
  * This file is part of NER's PM Dashboard and licensed under GNU AGPLv3.
  * See the LICENSE file in the repository root folder for details.
  */
-
+
