+++ src/components/projects/create-wp-form/create-wp-form.tsx
@@ -3,7 +3,7 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useHistory } from "react-router-dom";
 import CreateWPFormView from "./create-wp-form/create-wp-form";
