+++ src/components/shared/editable-text-input-list/editable-text-input-list.test.tsx
@@ -3,7 +3,7 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import { render, screen } from "../../../../../test-support/test-utils";
+import { render, screen } from "../../../test-support/test-utils";
 import EditableTextInputList from "./editable-text-input-list";
 
 const mockItems = ["tee hee", "yahello", "yeet", "yoink"];
